requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
psycopg2-binary>=2.9.0
configparser>=5.3.0

# Deep Learning
torch>=2.0.0
torch-geometric>=2.3.0
networkx>=3.1.0

# Data processing
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# File processing
ijson>=3.2.0

# Development
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0

# Optional: GPU support
# torch-cuda>=2.0.0  # Uncomment for CUDA support